composition-rel reform: - apply changes developed by Michael Steigman for making rel-types   configurable whether these should be composable or not   (default is: these are composable, which was up to now the only   choice) - improved documentation of log argument lists of SQL functions - Background:   http://openacs.org/forums/message-view?message_id=5330734   http://openacs.org/forums/message-view?message_id=4031049 - bump version number of acs-kernel to 5.9.1d16

+++ sql/postgresql/evaluation-create.sql
@@ -207,7 +207,7 @@ select acs_rel_type__create_type(
         'evaluation_task_groups',
         null,
         0,
-        null,
+        null::integer,
         'user',
         null,
         0,
